Move UI files to client folder
Move components, pages, HOCs and containers directories inside client directory.
