+++ babel.config.js
@@ -13,8 +13,8 @@ module.exports = {
         root: './',
         alias: {
           lib: './lib',
-          components: './imports/ui/components',
-          pages: './imports/ui/pages',
+          components: './client/components',
+          pages: './client/pages',
         },
       },
     ],
