+++ client/containers/dashboard.jsx
@@ -1,7 +1,7 @@
 import { Meteor } from 'meteor/meteor';
 import { withTracker } from 'meteor/react-meteor-data';
 import Entries from '/imports/api/entries/entries.js';
-import Dashboard from '/imports/ui/pages/dashboard';
+import Dashboard from '/client/pages/dashboard';
 
 // eslint-disable-next-line no-undef
 const categoryId = new ReactiveVar(3);
