+++ client/hoc/AuthenticatedRoute.jsx
