+++ client/pages/dashboard/dashboard_matches_section/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { PropTypes } from 'prop-types';
 
 import { Card, Nav, Tab } from 'react-bootstrap';
-import MatchCard from '/imports/ui/components/match_card';
+import MatchCard from '/client/components/match_card';
 
 const MatchesSection = props => {
   const { ownEntries } = props;
