+++ client/pages/faq.jsx
