+++ client/pages/set_password.jsx
